Adding condition to copy zpool.cache

+++ src-sh/pcbsd-utils/pc-sysinstall/backend/functions-unmount.sh
@@ -237,7 +237,11 @@ setup_grub()
   if [ ! -d "${FSMNT}/boot/zfs/" ] ; then 
      rc_halt "mkdir ${FSMNT}/boot/zfs"
   fi
-  rc_halt "cp /boot/zfs/zpool.cache ${FSMNT}/boot/zfs/"
+  
+  # GhostBSD doesn't use ZFS.
+  if [ -e "/boot/zfs/zpool.cache"] ; then
+    rc_halt "cp /boot/zfs/zpool.cache ${FSMNT}/boot/zfs/"
+  fi
 
   if [ ! -e "${FSMNT}/boot/kernel/zfs" ] ; then
     rc_halt "ln -s ../zfs ${FSMNT}/boot/kernel/zfs"
